Make chunk_bytes a usize across all table constructors
Per review: BlobTable::create, GeometryTable::create, CoordTable::create,
GraphTable::create, U64Set::create, and StringCounts::create now take
chunk_bytes: usize instead of u64, matching EXTERNAL_SORT_CHUNK_BYTES's
own type -- we'll never run on a 32-bit machine, so there's no precision
concern, and it reads more naturally at call sites (no as u64 noise
passing EXTERNAL_SORT_CHUNK_BYTES or EXTERNAL_SORT_CHUNK_BYTES /
CONCURRENT_SORTS through). The two call sites that build a
MemoryLimitedBufferBuilder directly (BlobTable, StringCounts) still cast
to u64 right at that call, since that's what the ext_sort crate's own
API requires; nothing else changes.

+++ src/atp/wikidata_ids.rs
@@ -66,7 +66,7 @@ pub fn collect_wikidata_ids(atp: &Path, workdir: &Path) -> Result<PathBuf> {
         ids.into_iter(),
         workdir,
         &out,
-        crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+        crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
     )?;
     log::info!(count = count; "collect_wikidata_ids: done");
     Ok(out)
